#WIP scoring function update + cleanup code base
