+++ src/main/scala/beam/agentsim/agents/household/FastHouseholdCAVScheduling.scala
@@ -1,19 +1,23 @@
 package beam.agentsim.agents.household
+import akka.actor.ActorRef
 import beam.agentsim.agents._
 import beam.agentsim.agents.planning.{BeamPlan, Trip}
 import beam.agentsim.agents.vehicles.VehicleProtocol.StreetVehicle
-import beam.agentsim.agents.vehicles.{BeamVehicle, BeamVehicleType}
+import beam.agentsim.agents.vehicles.{BeamVehicle, BeamVehicleType, VehiclePersonId}
 import beam.agentsim.events.SpaceTime
 import beam.router.BeamRouter.{EmbodyWithCurrentTravelTime, RoutingRequest}
-import beam.router.{BeamRouter, BeamSkimmer, RouteHistory}
 import beam.router.Modes.BeamMode
 import beam.router.Modes.BeamMode.CAV
+import beam.router.{BeamRouter, BeamSkimmer, Modes, RouteHistory}
 import beam.sim.BeamServices
+import beam.utils.logging.ExponentialLoggerWrapperImpl
 import com.conveyal.r5.transit.TransportNetwork
 import org.matsim.api.core.v01.Id
+import org.matsim.api.core.v01.population.{Leg, Person}
 import org.matsim.households.Household
 import org.matsim.vehicles.Vehicle
 
+import scala.collection.JavaConverters._
 import scala.collection.immutable.{List, Map}
 import scala.collection.mutable
 import scala.util.control.Breaks._
@@ -28,48 +32,37 @@ class FastHouseholdCAVScheduling(
   val beamServices: Option[BeamServices] = None
 )(implicit val population: org.matsim.api.core.v01.population.Population) {
 
-  import beam.agentsim.agents.memberships.Memberships.RankedGroup._
-
   import scala.collection.mutable.{ListBuffer => MListBuffer}
-  private val householdPlans = household.members
-    .take(limitCavToXPersons)
-    .map(
-      person => BeamPlan(person.getSelectedPlan)
-    )
 
-  // *******
-  // Convert to CAV SCHEDULE
-  // *******
-  def getAllFeasibleCAVSchedules: List[List[CAVSchedule]] = {
-    val output = mutable.ListBuffer[List[CAVSchedule]]()
-    getAllFeasibleSchedules.foreach { x =>
-      output.append(x.schedulesMap.map(y => CAVSchedule(y._2, y._1, 0)).toList)
-    }
-    output.toList
+  def getKLowestSumOfDelaysSchedules(k: Int): List[List[CAVSchedule]] = {
+    getAllFeasibleSchedules
+      .sortBy(_.householdScheduleCost.sumOfDelays.foldLeft(0)(_ + _._2))
+      .groupBy(_.householdScheduleCost.sumOfDelays.foldLeft(0)(_ + _._2))
+      .mapValues(_.maxBy(_.schedulesMap.foldLeft(0)((a, b) => a + b._2.schedule.size)))
+      .toList
+      .sortBy(_._1)
+      .map(_._2)
+      .take(k)
+      .map(_.schedulesMap.values.toList)
   }
-
-  def getKBestCAVSchedules(k: Int): List[CAVSchedule] = {
-    getKBestSchedules(k).headOption
-      .map(_.schedulesMap.map(y => CAVSchedule(y._2, y._1, 0)).toList)
-      .getOrElse(List.empty)
-  }
-
-  def getBestCAVScheduleWithLongestChain: List[CAVSchedule] = {
-    getBestScheduleWithLongestChain.headOption
-      .map(_.schedulesMap.map(y => CAVSchedule(y._2, y._1, 0)).toList)
-      .getOrElse(List.empty)
-  }
-  // *******
 
   // get k lowest scored schedules
-  def getKBestSchedules(k: Int): List[HouseholdSchedule] = {
-    getAllFeasibleSchedules.sortBy(_.householdScheduleCost.totalTravelTime).take(k)
+  def getKBestSchedules(k: Int): List[List[CAVSchedule]] = {
+    getAllFeasibleSchedules
+      .sortBy(_.householdScheduleCost.totalTravelTime)
+      .groupBy(_.householdScheduleCost.totalTravelTime)
+      .mapValues(_.maxBy(_.schedulesMap.foldLeft(0)((a, b) => a + b._2.schedule.size)))
+      .toList
+      .sortBy(_._1)
+      .map(_._2)
+      .take(k)
+      .map(_.schedulesMap.values.toList)
   }
 
   // ***
-  def getBestScheduleWithLongestChain: List[HouseholdSchedule] = {
-    val mapRank = getAllFeasibleSchedules.map(x => x -> x.schedulesMap.foldLeft(0)((a, b) => a + b._2.size)).toMap
-
+  def getBestProductiveSchedule: List[CAVSchedule] = {
+    val mapRank =
+      getAllFeasibleSchedules.map(x => x -> x.schedulesMap.foldLeft(0)((a, b) => a + b._2.schedule.size)).toMap
     var output = List.empty[HouseholdSchedule]
     if (mapRank.nonEmpty) {
       val maxRank = mapRank.maxBy(_._2)._2
@@ -80,47 +73,53 @@ class FastHouseholdCAVScheduling(
         .sortBy(_.householdScheduleCost.totalTravelTime)
         .take(1)
     }
-    output
+    output.headOption.map(_.schedulesMap.values.toList).getOrElse(List.empty)
   }
 
   def getAllFeasibleSchedules: List[HouseholdSchedule] = {
-    val householdTrips = HouseholdTrips.get(householdPlans, householdVehicles.size, timeWindow, skimmer)
-    val cavVehicles = householdVehicles.filter(_.beamVehicleType.automationLevel > 3)
-    val householdSchedules = mutable.ListBuffer.empty[HouseholdSchedule]
-    if (householdTrips.nonEmpty && cavVehicles.nonEmpty) {
-      val emptySchedule =
-        HouseholdSchedule(cavVehicles.map(_ -> List(householdTrips.get.homePickup)).toMap, householdTrips.get)
-      householdSchedules.append(emptySchedule)
-      breakable {
-        householdTrips.get.requests.foreach { requests =>
-          val HouseholdSchedulesToAppend = mutable.ListBuffer.empty[HouseholdSchedule]
-          val HouseholdSchedulesToDelete = mutable.ListBuffer.empty[HouseholdSchedule]
-          householdSchedules.foreach { hhSchedule =>
-            val newHouseholdSchedules = hhSchedule.check(requests)
-            HouseholdSchedulesToAppend.appendAll(newHouseholdSchedules)
-          }
-          householdSchedules.--=(HouseholdSchedulesToDelete)
-          householdSchedules.appendAll(HouseholdSchedulesToAppend)
-          if (householdSchedules.size >= stopSearchAfterXSolutions) {
-            break
+    HouseholdTrips.get(household, householdVehicles, limitCavToXPersons, householdVehicles.size, timeWindow, skimmer) match {
+      case Some(householdTrips) if householdTrips.cavVehicles.nonEmpty =>
+        val householdSchedules = mutable.ListBuffer.empty[HouseholdSchedule]
+        val emptySchedule =
+          HouseholdSchedule(
+            householdTrips.cavVehicles.map(cav => cav -> CAVSchedule(List(householdTrips.homePickup), cav, 0)).toMap,
+            householdTrips
+          )
+        householdSchedules.append(emptySchedule)
+        breakable {
+          householdTrips.requests.foreach { requests =>
+            val HouseholdSchedulesToAppend = mutable.ListBuffer.empty[HouseholdSchedule]
+            val HouseholdSchedulesToDelete = mutable.ListBuffer.empty[HouseholdSchedule]
+            householdSchedules.foreach { hhSchedule =>
+              val newHouseholdSchedules = hhSchedule.check(requests)
+              HouseholdSchedulesToAppend.appendAll(newHouseholdSchedules)
+            }
+            householdSchedules.--=(HouseholdSchedulesToDelete)
+            householdSchedules.appendAll(HouseholdSchedulesToAppend)
+            if (householdSchedules.size >= stopSearchAfterXSolutions) {
+              break
+            }
           }
         }
-      }
-      householdSchedules.-=(emptySchedule)
+        householdSchedules.-=(emptySchedule)
+//        beam.sandbox.CavRun
+//          .perfSchedules(householdSchedules.map(x => (x.schedulesMap.values.toList, x.householdScheduleCost)).toList)
+        householdSchedules.toList
+      case _ => List.empty[HouseholdSchedule]
     }
-    householdSchedules.toList
   }
 
   case class HouseholdSchedule(
-    schedulesMap: Map[BeamVehicle, List[MobilityRequest]],
+    schedulesMap: Map[BeamVehicle, CAVSchedule],
     householdScheduleCost: HouseholdTrips
   ) {
 
     def check(requests: List[MobilityRequest]): List[HouseholdSchedule] = {
       val outHouseholdSchedule = MListBuffer.empty[HouseholdSchedule]
       breakable {
-        for ((cav, schedule) <- schedulesMap.toArray.sortBy(_._2.size)(Ordering[Int].reverse)) {
-          getScheduleOrNone(cav, schedule ++ requests, timeWindow) match {
+        for ((cav, cavSchedule) <- schedulesMap.toArray.sortBy(_._2.schedule.size)(Ordering[Int].reverse)) {
+          // prioritizing CAVs with high usage
+          getScheduleOrNone(cav, cavSchedule, requests, timeWindow) match {
             case Some(s) =>
               outHouseholdSchedule.append(s)
               break
@@ -133,19 +132,20 @@ class FastHouseholdCAVScheduling(
 
     private def getScheduleOrNone(
       cav: BeamVehicle,
+      cavSchedule: CAVSchedule,
       requests: List[MobilityRequest],
       timeWindow: Map[MobilityRequestTrait, Int]
     ): Option[HouseholdSchedule] = {
-      val sortedRequests = requests.filter(_.tag != Relocation).sortBy(_.time)
-      val startRequest = sortedRequests.head
-      var occupancy = 0
+      if (cavSchedule.occupancy > cav.beamVehicleType.seatingCapacity)
+        return None
 
+      val sortedRequests = (cavSchedule.schedule ++ requests).filter(_.tag != Relocation).sortBy(_.time)
+      val startRequest = sortedRequests.head
       val newHouseholdSchedule = MListBuffer(startRequest.copy())
       var newHouseholdScheduleCost = householdScheduleCost.copy()
+      var newOccupancy: Int = cavSchedule.occupancy
 
       sortedRequests.drop(1).foreach { curReq =>
-        if (occupancy > cav.beamVehicleType.seatingCapacity)
-          return None
         val prevReq = newHouseholdSchedule.last
         val metric = skimmer.getTimeDistanceAndCost(
           prevReq.activity.getCoord,
@@ -154,11 +154,11 @@ class FastHouseholdCAVScheduling(
           BeamMode.CAR,
           BeamVehicleType.defaultCarBeamVehicleType.id
         )
-        var serviceTime = prevReq.serviceTime + metric.time.toInt
+        var serviceTime = prevReq.serviceTime + metric.time
         val ubTime = curReq.time + timeWindow(curReq.tag)
         val lbTime = curReq.time - timeWindow(curReq.tag)
         if (curReq.isPickup) {
-          if (serviceTime > ubTime || (occupancy != 0 && serviceTime < lbTime))
+          if (serviceTime > ubTime || (newOccupancy != 0 && serviceTime < lbTime))
             return None
           else if (serviceTime >= lbTime && serviceTime <= ubTime) {
             serviceTime = if (serviceTime < curReq.time) curReq.time else serviceTime
@@ -173,109 +173,52 @@ class FastHouseholdCAVScheduling(
             newHouseholdSchedule.append(relocationRequest)
             serviceTime = curReq.time
           }
-          newHouseholdSchedule.append(curReq.copy(serviceTime = serviceTime))
-          occupancy += 1
+          newOccupancy += 1
+          newHouseholdSchedule.append(curReq.copy(serviceTime = serviceTime, vehicleOccupancy = Some(newOccupancy)))
         } else if (curReq.isDropoff) {
           if (serviceTime < lbTime || serviceTime > ubTime) {
             return None
           }
-          val pickupReq = newHouseholdSchedule.filter(x => x.isPickup && x.person == curReq.person).last
-          val cavTripTravelTime = serviceTime - pickupReq.time // it includes the waiting time
+          val index = newHouseholdSchedule.lastIndexWhere(_.trip == curReq.trip)
+          val pickupReq = newHouseholdSchedule.apply(index)
+          newOccupancy -= 1
+          newHouseholdSchedule.append(
+            curReq
+              .copy(serviceTime = serviceTime, pickupRequest = Some(pickupReq), vehicleOccupancy = Some(newOccupancy))
+          )
+          // it includes the waiting time
+          val cavTripTravelTime = computeSharedTravelTime(newHouseholdSchedule.slice(index, newHouseholdSchedule.size))
           val newTotalTravelTime = newHouseholdScheduleCost.totalTravelTime -
           newHouseholdScheduleCost.tripTravelTime(curReq.trip) + cavTripTravelTime
-          if (newTotalTravelTime > newHouseholdScheduleCost.defaultTravelTime)
+          if (newTotalTravelTime > newHouseholdScheduleCost.baseTotalTravelTime)
             return None
+          val sumOfDelays = (pickupReq.serviceTime - pickupReq.time) + (serviceTime - curReq.time)
+          if (sumOfDelays < 0) {
+            println("neg")
+          }
           newHouseholdScheduleCost = newHouseholdScheduleCost.copy(
-            tripTravelTime = newHouseholdScheduleCost.tripTravelTime + (pickupReq.trip -> cavTripTravelTime),
-            totalTravelTime = newTotalTravelTime
+            tripTravelTime = newHouseholdScheduleCost.tripTravelTime + (curReq.trip -> cavTripTravelTime),
+            totalTravelTime = newTotalTravelTime,
+            sumOfDelays = newHouseholdScheduleCost.sumOfDelays +
+            (curReq.person.get.personId -> (sumOfDelays + newHouseholdScheduleCost
+              .sumOfDelays(curReq.person.get.personId)))
           )
-          newHouseholdSchedule.append(curReq.copy(serviceTime = serviceTime, pickupRequest = Some(pickupReq)))
-          occupancy -= 1
         }
       }
-
       Some(
         HouseholdSchedule(
-          this.schedulesMap.filterKeys(_ != cav) + (cav -> newHouseholdSchedule.toList),
+          this.schedulesMap.filterKeys(_ != cav) + (cav -> CAVSchedule(newHouseholdSchedule.toList, cav, newOccupancy)),
           newHouseholdScheduleCost
         )
       )
     }
 
-    case class RouteOrEmbodyRequest(routeReq: Option[RoutingRequest], embodyReq: Option[EmbodyWithCurrentTravelTime])
-
-    def toRoutingRequests(
-      beamServices: BeamServices,
-      transportNetwork: TransportNetwork,
-      routeHistory: RouteHistory,
-      cav: BeamVehicle
-    ): (List[Option[RouteOrEmbodyRequest]], List[MobilityRequest]) = {
-      var newMobilityRequests = List[MobilityRequest]()
-      val requestList = (schedulesMap(cav).reverse :+ schedulesMap(cav).head).tail
-        .sliding(2)
-        .map { wayPoints =>
-          val orig = wayPoints(0)
-          val dest = wayPoints(1)
-          val origin = SpaceTime(orig.activity.getCoord, Math.round(orig.time))
-          if (beamServices.geo.distUTMInMeters(orig.activity.getCoord, dest.activity.getCoord) < beamServices.beamConfig.beam.agentsim.thresholdForWalkingInMeters) {
-            newMobilityRequests = newMobilityRequests :+ orig
-            None
-          } else {
-            val theVehicle = StreetVehicle(
-              Id.create(cav.id.toString, classOf[Vehicle]),
-              cav.beamVehicleType.id,
-              origin,
-              CAV,
-              asDriver = true
-            )
-            val origLink = beamServices.geo.getNearestR5Edge(
-              transportNetwork.streetLayer,
-              beamServices.geo.utm2Wgs(orig.activity.getCoord),
-              10E3
-            )
-            val destLink = beamServices.geo.getNearestR5Edge(
-              transportNetwork.streetLayer,
-              beamServices.geo.utm2Wgs(dest.activity.getCoord),
-              10E3
-            )
-            routeHistory.getRoute(origLink, destLink, orig.time) match {
-              case Some(rememberedRoute) =>
-                val embodyReq = BeamRouter.linkIdsToEmbodyRequest(
-                  rememberedRoute,
-                  theVehicle,
-                  origin.time,
-                  CAV,
-                  beamServices,
-                  orig.activity.getCoord,
-                  dest.activity.getCoord
-                )
-                newMobilityRequests = newMobilityRequests :+ orig.copy(routingRequestId = Some(embodyReq.requestId))
-                Some(RouteOrEmbodyRequest(None, Some(embodyReq)))
-              case None =>
-                val routingRequest = RoutingRequest(
-                  orig.activity.getCoord,
-                  dest.activity.getCoord,
-                  origin.time,
-                  IndexedSeq(),
-                  IndexedSeq(
-                    StreetVehicle(
-                      Id.create(cav.id.toString, classOf[Vehicle]),
-                      cav.beamVehicleType.id,
-                      origin,
-                      CAV,
-                      asDriver = true
-                    )
-                  )
-                )
-                newMobilityRequests = newMobilityRequests :+ orig.copy(
-                  routingRequestId = Some(routingRequest.requestId)
-                )
-                Some(RouteOrEmbodyRequest(Some(routingRequest), None))
-            }
-          }
-        }
-        .toList
-      (requestList, newMobilityRequests)
+    private def computeSharedTravelTime(requestsSeq: MListBuffer[MobilityRequest]): Int = {
+      val waitTime = requestsSeq.head.serviceTime - requestsSeq.head.time
+      requestsSeq.filter(x => x.isPickup || x.isDropoff).sliding(2).foldLeft(waitTime) {
+        case (acc, Seq(prevReq, nextReq)) =>
+          acc + ((nextReq.serviceTime - prevReq.serviceTime) / prevReq.vehicleOccupancy.getOrElse(1))
+      }
     }
 
     override def toString: String = {
@@ -283,7 +226,7 @@ class FastHouseholdCAVScheduling(
         .foldLeft(new StringBuilder) {
           case (output, (cav, schedules)) =>
             output.append(s"cavid: ${cav.id}\n")
-            val outputBis = schedules.foldLeft(output) {
+            val outputBis = schedules.schedule.foldLeft(output) {
               case (outputBisBis, schedule) =>
                 outputBisBis.append(s"\t$schedule\n")
             }
@@ -291,47 +234,248 @@ class FastHouseholdCAVScheduling(
         }
         .insert(
           0,
-          s"\nhid:None | tt:${householdScheduleCost.totalTravelTime} | tt0:${householdScheduleCost.defaultTravelTime}.\n"
+          s"\nhid:None | tt:${householdScheduleCost.totalTravelTime} | base-tt:${householdScheduleCost.baseTotalTravelTime}.\n"
         )
         .toString
     }
   }
 }
 
+case class CAVSchedule(schedule: List[MobilityRequest], cav: BeamVehicle, occupancy: Int) {
+  case class RouteOrEmbodyRequest(routeReq: Option[RoutingRequest], embodyReq: Option[EmbodyWithCurrentTravelTime])
+
+  def toRoutingRequests(
+    beamServices: BeamServices,
+    transportNetwork: TransportNetwork,
+    routeHistory: RouteHistory
+  ): (List[Option[RouteOrEmbodyRequest]], List[MobilityRequest]) = {
+    var newMobilityRequests = List[MobilityRequest]()
+    val requestList = (schedule.reverse :+ schedule.head).tail
+      .sliding(2)
+      .map { wayPoints =>
+        val orig = wayPoints(0)
+        val dest = wayPoints(1)
+        val origin = SpaceTime(orig.activity.getCoord, Math.round(orig.time))
+        if (beamServices.geo.distUTMInMeters(orig.activity.getCoord, dest.activity.getCoord) < beamServices.beamConfig.beam.agentsim.thresholdForWalkingInMeters) {
+          newMobilityRequests = newMobilityRequests :+ orig
+          None
+        } else {
+          val theVehicle = StreetVehicle(
+            Id.create(cav.id.toString, classOf[Vehicle]),
+            cav.beamVehicleType.id,
+            origin,
+            CAV,
+            asDriver = true
+          )
+          val origLink = beamServices.geo.getNearestR5Edge(
+            transportNetwork.streetLayer,
+            beamServices.geo.utm2Wgs(orig.activity.getCoord),
+            10E3
+          )
+          val destLink = beamServices.geo.getNearestR5Edge(
+            transportNetwork.streetLayer,
+            beamServices.geo.utm2Wgs(dest.activity.getCoord),
+            10E3
+          )
+          routeHistory.getRoute(origLink, destLink, orig.time) match {
+            case Some(rememberedRoute) =>
+              val embodyReq = BeamRouter.linkIdsToEmbodyRequest(
+                rememberedRoute,
+                theVehicle,
+                origin.time,
+                CAV,
+                beamServices,
+                orig.activity.getCoord,
+                dest.activity.getCoord
+              )
+              newMobilityRequests = newMobilityRequests :+ orig.copy(routingRequestId = Some(embodyReq.requestId))
+              Some(RouteOrEmbodyRequest(None, Some(embodyReq)))
+            case None =>
+              val routingRequest = RoutingRequest(
+                orig.activity.getCoord,
+                dest.activity.getCoord,
+                origin.time,
+                IndexedSeq(),
+                IndexedSeq(
+                  StreetVehicle(
+                    Id.create(cav.id.toString, classOf[Vehicle]),
+                    cav.beamVehicleType.id,
+                    origin,
+                    CAV,
+                    asDriver = true
+                  )
+                )
+              )
+              newMobilityRequests = newMobilityRequests :+ orig.copy(
+                routingRequestId = Some(routingRequest.requestId)
+              )
+              Some(RouteOrEmbodyRequest(Some(routingRequest), None))
+          }
+        }
+      }
+      .toList
+    (requestList, newMobilityRequests)
+  }
+
+}
+
 case class HouseholdTrips(
+  household: Household,
   requests: List[List[MobilityRequest]],
+  cavVehicles: List[BeamVehicle],
   homePickup: MobilityRequest,
-  defaultTravelTime: Int,
+  baseTotalTravelTime: Int,
   tripTravelTime: Map[Trip, Int],
-  totalTravelTime: Int
+  totalTravelTime: Int,
+  sumOfDelays: Map[Id[Person], Int]
 ) {
   override def toString: String = requests.toString
-
 }
 
 object HouseholdTrips {
 
   def get(
-    householdPlans: Seq[BeamPlan],
+    household: Household,
+    householdVehicles: List[BeamVehicle],
+    limitCavToXPersons: Int,
     householdNbOfVehicles: Int,
     timeWindow: Map[MobilityRequestTrait, Int],
     skim: BeamSkimmer
-  ): Option[HouseholdTrips] = {
+  )(implicit population: org.matsim.api.core.v01.population.Population): Option[HouseholdTrips] = {
+    import beam.agentsim.agents.memberships.Memberships.RankedGroup._
+    val householdPlans = household.members
+      .take(limitCavToXPersons)
+      .map(
+        person => BeamPlan(person.getSelectedPlan)
+      )
+    val cavVehicles = householdVehicles.filter(_.beamVehicleType.automationLevel > 3)
     val (requests, firstPickupOfTheDay, tripTravelTime, totTravelTime) =
       HouseholdTripsHelper.getListOfPickupsDropoffs(householdPlans, householdNbOfVehicles, skim)
-    // Sum(tDi - tPi) <= Sum(tauDi - tauPi) + (alpha + beta)|R|/2
-    val sumTimeWindows = requests.size * timeWindow.foldLeft(0)(_ + _._2)
-
-    // adding a time window to the total travel time
     firstPickupOfTheDay map (
       homePickup =>
         HouseholdTrips(
+          household,
           requests,
+          cavVehicles,
           homePickup.copy(person = None, tag = Init),
-          totTravelTime + sumTimeWindows,
+          totTravelTime,
           tripTravelTime.toMap,
-          totTravelTime
+          totTravelTime,
+          household.getMemberIds.asScala.map(_ -> 0).toMap
         )
     )
+  }
+}
+
+// Helper classes for convert Beam plans to MobilityServiceRequest
+case class HouseholdTripsException(message: String, cause: Throwable = null) extends Exception(message, cause)
+case class HouseholdTripsLogger(name: String) extends ExponentialLoggerWrapperImpl(name)
+
+object HouseholdTripsHelper {
+
+  import scala.collection.mutable.{ListBuffer => MListBuffer, Map => MMap}
+  import scala.util.control.Breaks._
+  val logger = HouseholdTripsLogger(getClass.getName)
+
+  def getDefaultMode(legOption: Option[Leg], nbVehicles: Int): BeamMode = {
+    legOption
+      .flatMap(leg => BeamMode.fromString(leg.getMode))
+      .getOrElse(if (nbVehicles <= 0) BeamMode.TRANSIT else BeamMode.CAR)
+  }
+
+  def getListOfPickupsDropoffs(
+    householdPlans: Seq[BeamPlan],
+    householdNbOfVehicles: Int,
+    skim: BeamSkimmer
+  ): (List[List[MobilityRequest]], Option[MobilityRequest], MMap[Trip, Int], Int) = {
+    val requests = MListBuffer.empty[List[MobilityRequest]]
+    val tours = MListBuffer.empty[MobilityRequest]
+    val tripTravelTime = MMap[Trip, Int]()
+    var totTravelTime = 0
+    var firstPickupOfTheDay: Option[MobilityRequest] = None
+    breakable {
+      householdPlans.foldLeft(householdNbOfVehicles) {
+        case (counter, plan) =>
+          val usedCarOut = plan.trips.sliding(2).foldLeft(false) {
+            case (usedCar, Seq(prevTrip, curTrip)) =>
+              val (pickup, dropoff, travelTime) = getPickupAndDropoff(plan, curTrip, prevTrip, counter, skim)
+              if (firstPickupOfTheDay.isEmpty || firstPickupOfTheDay.get.time > pickup.time)
+                firstPickupOfTheDay = Some(pickup)
+              tours.append(pickup)
+              tours.append(dropoff)
+              if (!Modes.isChainBasedMode(pickup.defaultMode) || tours.head.trip.parentTour != pickup.trip.parentTour) {
+                requests.append(tours.toList)
+                tours.clear()
+              }
+              tripTravelTime(curTrip) = travelTime
+              totTravelTime += travelTime
+              if (pickup.defaultMode == BeamMode.CAR) true else usedCar
+          }
+          requests.append(tours.toList)
+          tours.clear()
+          if (usedCarOut) counter - 1 else counter
+      }
+    }
+    (requests.toList, firstPickupOfTheDay, tripTravelTime, totTravelTime)
+  }
+
+  def getPickupAndDropoff(
+    plan: BeamPlan,
+    curTrip: Trip,
+    prevTrip: Trip,
+    counter: Int,
+    skim: BeamSkimmer
+  ): (MobilityRequest, MobilityRequest, Int) = {
+    val legTrip = curTrip.leg
+    val defaultMode = getDefaultMode(legTrip, counter)
+    val travelTime = skim
+      .getTimeDistanceAndCost(
+        prevTrip.activity.getCoord,
+        curTrip.activity.getCoord,
+        0,
+        defaultMode,
+        org.matsim.api.core.v01.Id.create[BeamVehicleType]("", classOf[BeamVehicleType])
+      )
+      .time
+
+    val startTime = prevTrip.activity.getEndTime.toInt
+    val arrivalTime = startTime + travelTime
+
+    val nextTripStartTime = curTrip.activity.getEndTime
+    if (nextTripStartTime != Double.NegativeInfinity && startTime >= nextTripStartTime.toInt) {
+      logger.warn(
+        s"Illegal plan for person ${plan.getPerson.getId.toString}, activity ends at $startTime which is later than the next activity ending at $nextTripStartTime"
+      )
+      break
+    } else if (nextTripStartTime != Double.NegativeInfinity && arrivalTime > nextTripStartTime.toInt) {
+      logger.warn(
+        "The necessary travel time to arrive to the next activity is beyond the end time of the same activity"
+      )
+      break
+    }
+
+    val vehiclePersonId =
+      VehiclePersonId(Id.create(plan.getPerson.getId, classOf[Vehicle]), plan.getPerson.getId, ActorRef.noSender)
+
+    val pickup = MobilityRequest(
+      Some(vehiclePersonId),
+      prevTrip.activity,
+      startTime,
+      curTrip,
+      defaultMode,
+      Pickup,
+      startTime
+    )
+    val dropoff = MobilityRequest(
+      Some(vehiclePersonId),
+      curTrip.activity,
+      arrivalTime,
+      curTrip,
+      defaultMode,
+      Dropoff,
+      arrivalTime,
+      pickupRequest = Some(pickup)
+    )
+    (pickup, dropoff, travelTime)
   }
 }
