+++ src/test/scala/beam/agentsim/agents/household/FastHouseholdCAVSchedulingSpec.scala
@@ -7,7 +7,8 @@ import akka.util.Timeout
 import beam.agentsim.agents.choice.mode.ModeIncentive
 import beam.agentsim.agents.choice.mode.ModeIncentive.Incentive
 import beam.agentsim.agents.vehicles.EnergyEconomyAttributes.Powertrain
-import beam.agentsim.agents.vehicles.FuelType.FuelType
+import beam.agentsim.agents.vehicles.FuelType.{FuelType, Gasoline}
+import beam.agentsim.agents.vehicles.VehicleCategory.Car
 import beam.agentsim.agents.vehicles.{BeamVehicle, BeamVehicleType}
 import beam.agentsim.agents.{Dropoff, Pickup}
 import beam.agentsim.infrastructure.TAZTreeMap
@@ -19,15 +20,20 @@ import beam.sim.config.{BeamConfig, MatSimBeamConfigBuilder}
 import beam.utils.TestConfigUtils.testConfig
 import com.typesafe.config.ConfigFactory
 import org.matsim.api.core.v01.network.Network
-import org.matsim.api.core.v01.population.Population
-import org.matsim.api.core.v01.{Id, Scenario}
+import org.matsim.api.core.v01.population.{Activity, Person, Plan, Population}
+import org.matsim.api.core.v01.{Coord, Id, Scenario}
+import org.matsim.core.config.ConfigUtils
 import org.matsim.core.controler.MatsimServices
-import org.matsim.households.HouseholdsFactoryImpl
+import org.matsim.core.population.PopulationUtils
+import org.matsim.core.population.io.PopulationReader
+import org.matsim.core.scenario.ScenarioUtils
+import org.matsim.households.{Household, HouseholdImpl, HouseholdsFactoryImpl, HouseholdsReaderV10}
 import org.mockito.Mockito.{when, withSettings}
 import org.scalatest.mockito.MockitoSugar
 import org.scalatest.{BeforeAndAfterAll, FunSpecLike, Matchers}
 
 import scala.collection.immutable.List
+import scala.collection.{mutable, JavaConverters}
 import scala.concurrent.ExecutionContext
 
 class FastHouseholdCAVSchedulingSpec
@@ -81,15 +87,14 @@ class FastHouseholdCAVSchedulingSpec
         new BeamVehicle(
           Id.createVehicleId("id1"),
           new Powertrain(0.0),
-          HouseholdCAVSchedulingTest.defaultCAVBeamVehicleType
+          FastHouseholdCAVSchedulingSpec.defaultCAVBeamVehicleType
         )
       )
-      val (pop: Population, household) = HouseholdCAVSchedulingTest.scenario1(cavs)
-
+      val (pop: Population, household) = FastHouseholdCAVSchedulingSpec.scenario1(cavs)
       val alg = new FastHouseholdCAVScheduling(household, cavs, Map((Pickup, 2), (Dropoff, 2)), skimmer = skimmer)(pop)
       val schedules = alg.getAllFeasibleSchedules
       schedules should have length 1
-      schedules foreach (_.schedulesMap(cavs.head) should (have length 1 or have length 6))
+      schedules foreach (_.schedulesMap(cavs.head).schedule should have length 6)
       println(s"*** scenario 1 *** ${schedules.size} combinations")
     }
 
@@ -98,11 +103,15 @@ class FastHouseholdCAVSchedulingSpec
         new BeamVehicle(
           Id.createVehicleId("id1"),
           new Powertrain(0.0),
-          HouseholdCAVSchedulingTest.defaultCAVBeamVehicleType
+          FastHouseholdCAVSchedulingSpec.defaultCAVBeamVehicleType
         ),
-        new BeamVehicle(Id.createVehicleId("id2"), new Powertrain(0.0), BeamVehicleType.defaultCarBeamVehicleType)
+        new BeamVehicle(
+          Id.createVehicleId("id2"),
+          new Powertrain(0.0),
+          BeamVehicleType.defaultCarBeamVehicleType
+        )
       )
-      val (pop: Population, household) = HouseholdCAVSchedulingTest.scenario2(cavs)
+      val (pop: Population, household) = FastHouseholdCAVSchedulingSpec.scenario2(cavs)
       val alg = new FastHouseholdCAVScheduling(
         household,
         cavs,
@@ -112,7 +121,7 @@ class FastHouseholdCAVSchedulingSpec
       )(pop)
       val schedules = alg.getAllFeasibleSchedules
       schedules should have length 3
-      schedules foreach (_.schedulesMap(cavs.head) should (have length 1 or (have length 6 or have length 10)))
+      schedules foreach (_.schedulesMap(cavs.head).schedule should (have length 1 or (have length 6 or have length 10)))
       println(s"*** scenario 2 *** ${schedules.size} combinations")
     }
 
@@ -121,15 +130,15 @@ class FastHouseholdCAVSchedulingSpec
         new BeamVehicle(
           Id.createVehicleId("id1"),
           new Powertrain(0.0),
-          HouseholdCAVSchedulingTest.defaultCAVBeamVehicleType
+          FastHouseholdCAVSchedulingSpec.defaultCAVBeamVehicleType
         ),
         new BeamVehicle(
           Id.createVehicleId("id2"),
           new Powertrain(0.0),
-          HouseholdCAVSchedulingTest.defaultCAVBeamVehicleType
+          FastHouseholdCAVSchedulingSpec.defaultCAVBeamVehicleType
         )
       )
-      val (pop: Population, household) = HouseholdCAVSchedulingTest.scenario5(cavs)
+      val (pop: Population, household) = FastHouseholdCAVSchedulingSpec.scenario5(cavs)
       val alg = new FastHouseholdCAVScheduling(
         household,
         cavs,
@@ -140,28 +149,29 @@ class FastHouseholdCAVSchedulingSpec
       // first check
       val schedules1 = alg.getAllFeasibleSchedules
       schedules1 should have length 3
-      schedules1 foreach (_.schedulesMap(cavs.head) should (have length 1 or (have length 6 or have length 10)))
+      schedules1 foreach (_.schedulesMap(cavs.head).schedule should (have length 1 or (have length 6 or have length 10)))
       println(s"*** scenario 5 *** ${schedules1.size} combinations")
       // second check
-      val schedules2 = alg.getBestScheduleWithLongestChain
-      println(schedules2)
+      val schedules2 = alg.getBestProductiveSchedule
+      schedules2.foldLeft(0)(_+_.schedule.size) shouldBe 11
       // third check
       val schedules3 = alg.getKBestSchedules(1)
-      println(schedules3)
+      schedules3.head.foldLeft(0)(_+_.schedule.size) shouldBe 11
     }
 
     it("be scalable") {
       var sum = 0
       var count = 0
       val t0 = System.nanoTime()
-      val (pop: Population, households) = HouseholdCAVSchedulingTest.scenarioPerformance()
+      val (pop: Population, households) = FastHouseholdCAVSchedulingSpec.scenarioPerformance()
       for ((household, vehicles) <- households) {
         val alg =
           new FastHouseholdCAVScheduling(
             household,
             vehicles,
-            Map((Pickup, 15 * 60), (Dropoff, 15 * 60)),
-            stopSearchAfterXSolutions = Int.MaxValue,
+            Map((Pickup, 5 * 60), (Dropoff, 10 * 60)),
+            stopSearchAfterXSolutions = 1000,
+            limitCavToXPersons = Int.MaxValue,
             skimmer = skimmer
           )(pop)
         val schedules = alg.getAllFeasibleSchedules
@@ -175,5 +185,154 @@ class FastHouseholdCAVSchedulingSpec
       println(s"*** scenario 6 *** ${sum / count} avg combinations per household, $elapsed sec elapsed ")
     }
   }
+}
 
+object FastHouseholdCAVSchedulingSpec {
+
+  val defaultCAVBeamVehicleType = BeamVehicleType(
+    Id.create("CAV-TYPE-DEFAULT", classOf[BeamVehicleType]),
+    4,
+    0,
+    4.5,
+    Gasoline,
+    3656.0,
+    3655980000.0,
+    vehicleCategory = Car,
+    automationLevel = 5
+  )
+
+  def scenario1(vehicles: List[BeamVehicle]): (Population, Household) = {
+    val sc: org.matsim.api.core.v01.Scenario = ScenarioUtils.createMutableScenario(ConfigUtils.createConfig())
+    ScenarioUtils.loadScenario(sc)
+    val household = new HouseholdsFactoryImpl().createHousehold(Id.create("dummy_scenario1", classOf[Household]))
+    val popFactory = sc.getPopulation.getFactory
+
+    val p: Person = popFactory.createPerson(Id.createPersonId(household.getId + "_P1"))
+    val homeCoord = new Coord(0, 0)
+    val H11: Activity = popFactory.createActivityFromCoord("home", homeCoord)
+    H11.setEndTime(8 * 3600 + 30 * 60)
+    val W1: Activity = popFactory.createActivityFromCoord("work", new Coord(30, 0))
+    W1.setEndTime(17 * 3600)
+    val H12: Activity = popFactory.createActivityFromCoord("home", homeCoord)
+    val plan: Plan = popFactory.createPlan()
+    plan.setPerson(p)
+    plan.addActivity(H11)
+    plan.addActivity(W1)
+    plan.addActivity(H12)
+    p.addPlan(plan)
+    sc.getPopulation.addPerson(p)
+
+    household.setMemberIds(JavaConverters.bufferAsJavaList(mutable.Buffer(p.getId)))
+    household.setVehicleIds(JavaConverters.seqAsJavaList(vehicles.map(veh => veh.toStreetVehicle.id)))
+    (sc.getPopulation, household)
+  }
+
+  def scenario2(vehicles: List[BeamVehicle]): (Population, Household) = {
+    val sc: org.matsim.api.core.v01.Scenario = ScenarioUtils.createMutableScenario(ConfigUtils.createConfig())
+    ScenarioUtils.loadScenario(sc)
+    val pop = sc.getPopulation
+    val homeCoord = new Coord(0, 0)
+    val household = new HouseholdsFactoryImpl().createHousehold(Id.create("dummyHH_scenario2", classOf[Household]))
+
+    val P1: Person = pop.getFactory.createPerson(Id.createPersonId(household.getId + "_P1"))
+    val H11: Activity = PopulationUtils.createActivityFromCoord("home", homeCoord)
+    H11.setEndTime(9 * 3600)
+    val W1: Activity = PopulationUtils.createActivityFromCoord("work1", new Coord(24166, 13820))
+    W1.setEndTime(17 * 3600)
+    val H12: Activity = PopulationUtils.createActivityFromCoord("home", homeCoord)
+    val plan1: Plan = pop.getFactory.createPlan()
+    plan1.setPerson(P1)
+    plan1.addActivity(H11)
+    plan1.addActivity(W1)
+    plan1.addActivity(H12)
+    P1.addPlan(plan1)
+    pop.addPerson(P1)
+
+    val P2: Person = pop.getFactory.createPerson(Id.createPersonId(household.getId + "_P2"))
+    val H21: Activity = PopulationUtils.createActivityFromCoord("home", homeCoord)
+    H21.setEndTime(9 * 3600 + 5 * 60)
+    val W2: Activity = PopulationUtils.createActivityFromCoord("work2", new Coord(20835, 0))
+    W2.setEndTime(17 * 3600 + 5 * 60)
+    val H22: Activity = PopulationUtils.createActivityFromCoord("home", homeCoord)
+    val plan2: Plan = pop.getFactory.createPlan()
+    plan2.setPerson(P2)
+    plan2.addActivity(H21)
+    plan2.addActivity(W2)
+    plan2.addActivity(H22)
+    P2.addPlan(plan2)
+    pop.addPerson(P2)
+
+    household.setMemberIds(JavaConverters.bufferAsJavaList(mutable.Buffer(P1.getId, P2.getId)))
+    household.setVehicleIds(JavaConverters.seqAsJavaList(vehicles.map(veh => veh.toStreetVehicle.id)))
+    (sc.getPopulation, household)
+  }
+
+  def scenarioPerformance(): (Population, List[(Household, List[BeamVehicle])]) = {
+    val sc: org.matsim.api.core.v01.Scenario = ScenarioUtils.createMutableScenario(ConfigUtils.createConfig())
+    ScenarioUtils.loadScenario(sc)
+    new PopulationReader(sc).readFile("test/input/sf-light/sample/25k/population.xml.gz")
+    new HouseholdsReaderV10(sc.getHouseholds).readFile("test/input/sf-light/sample/25k/households.xml")
+    val households: mutable.ListBuffer[(Household, List[BeamVehicle])] =
+      mutable.ListBuffer.empty[(Household, List[BeamVehicle])]
+    import scala.collection.JavaConverters._
+    for (hh: Household <- sc.getHouseholds.getHouseholds.asScala.values) {
+      //hh.asInstanceOf[HouseholdImpl].setMemberIds(hh.getMemberIds)
+      val vehicles = List[BeamVehicle](
+        new BeamVehicle(
+          Id.createVehicleId(hh.getId.toString + "-veh1"),
+          new Powertrain(0.0),
+          FastHouseholdCAVSchedulingSpec.defaultCAVBeamVehicleType
+        ),
+        new BeamVehicle(
+          Id.createVehicleId(hh.getId.toString + "-veh2"),
+          new Powertrain(0.0),
+          FastHouseholdCAVSchedulingSpec.defaultCAVBeamVehicleType
+        )
+      )
+      hh.asInstanceOf[HouseholdImpl]
+        .setVehicleIds(JavaConverters.seqAsJavaList(vehicles.map(veh => veh.toStreetVehicle.id)))
+      households.append((hh.asInstanceOf[HouseholdImpl], vehicles))
+    }
+    (sc.getPopulation, households.toList)
+  }
+
+  def scenario5(vehicles: List[BeamVehicle]): (Population, Household) = {
+    val sc: org.matsim.api.core.v01.Scenario = ScenarioUtils.createMutableScenario(ConfigUtils.createConfig())
+    ScenarioUtils.loadScenario(sc)
+    val pop = sc.getPopulation
+    val homeCoord = new Coord(0, 0)
+    val household = new HouseholdsFactoryImpl().createHousehold(Id.create("dummyHH_scenario5", classOf[Household]))
+
+    val P1: Person = pop.getFactory.createPerson(Id.createPersonId(household.getId + "_P1"))
+    val H11: Activity = PopulationUtils.createActivityFromCoord("home", homeCoord)
+    H11.setEndTime(9 * 3600)
+    val W1: Activity = PopulationUtils.createActivityFromCoord("work1", new Coord(24166, 13820))
+    W1.setEndTime(17 * 3600)
+    val H12: Activity = PopulationUtils.createActivityFromCoord("home", homeCoord)
+    val plan1: Plan = pop.getFactory.createPlan()
+    plan1.setPerson(P1)
+    plan1.addActivity(H11)
+    plan1.addActivity(W1)
+    plan1.addActivity(H12)
+    P1.addPlan(plan1)
+    pop.addPerson(P1)
+
+    val P2: Person = pop.getFactory.createPerson(Id.createPersonId(household.getId + "_P2"))
+    val H21: Activity = PopulationUtils.createActivityFromCoord("home", homeCoord)
+    H21.setEndTime(9 * 3600 + 5 * 60)
+    val W2: Activity = PopulationUtils.createActivityFromCoord("work2", new Coord(20835, 0))
+    W2.setEndTime(18 * 3600)
+    val H22: Activity = PopulationUtils.createActivityFromCoord("home", homeCoord)
+    val plan2: Plan = pop.getFactory.createPlan()
+    plan2.setPerson(P2)
+    plan2.addActivity(H21)
+    plan2.addActivity(W2)
+    plan2.addActivity(H22)
+    P2.addPlan(plan2)
+    pop.addPerson(P2)
+
+    household.setMemberIds(JavaConverters.bufferAsJavaList(mutable.Buffer(P1.getId, P2.getId)))
+    household.setVehicleIds(JavaConverters.seqAsJavaList(vehicles.map(veh => veh.toStreetVehicle.id)))
+    (sc.getPopulation, household)
+  }
 }
